Update date.cc
Typing Mistakes.

+++ xapian-applications/omega/date.cc
@@ -166,7 +166,7 @@ date_range_filter(const string & date_start, const string & date_end,
 	    }
 	    secs = time_t(str_to_long) * 24 * 60 * 60;
 	} else {
-	    long str_to_longlong;
+	    long long str_to_longlong;
 	    errno = 0;
 	    str_to_longlong = strtoll(date_span.c_str(),NULL,0);
 	    if (errno == ERANGE) {
